% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{absmin.criteria}
\alias{absmin.criteria}
\title{Absmin Criteria}
\usage{
absmin.criteria(x)
}
\arguments{
\item{x}{a \code{matrix} or \code{Matrix}, initial factor loadings matrix for which the rotation criterian is to be optimized.}
}
\description{
Calculate the absmin criteria.
This is a helper function for \link{absmin}.
}
