% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{distance}
\alias{distance}
\title{Matrix Distance}
\usage{
distance(x, y, method = "euclidean")
}
\arguments{
\item{x, y}{\code{matrix} or \code{Matrix}, of the same number of rows. The columns of \code{x} and \code{y} will be scaled to unit length.}

\item{method}{distance measure, "maximum", "cosine", or "euclidean" are implemented.}
}
\value{
\code{numeric}, the distance between two matrices.
}
\description{
Matrix Distance
}
