% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epca-package.R
\docType{package}
\name{epca-package}
\alias{epca}
\alias{epca-package}
\title{epca: Exploratory Principal Component Analysis}
\description{
Exploratory Principal Component Analysis (PCA) for large-scale dataset, powered by sparse multivariate decomposition (SMD). 
    The epca features sparse PCA and its applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https:github.com/fchen365/epca}
}

}
\author{
\strong{Maintainer}: Fan Chen \email{fan.chen@wisc.edu}

}
\keyword{internal}
