% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{evalThres}
\alias{evalThres}
\title{evaluate soft-thresholding function given cut-off}
\usage{
evalThres(mat, cutoff, norm.fun = norm11)
}
\value{
$value and $matrix after soft-thresholding
}
\description{
evaluate soft-thresholding function given cut-off
}
