% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fftshift}
\alias{fftshift}
\title{ffshift}
\usage{
fftshift(x)
}
\arguments{
\item{x}{a \code{matrix} of image.}
}
\description{
Rearrange the matrix so that zero frequency component is in the middle of the matrix.
This is similar to the \code{fftshift} function in MATLAB
}
\references{
Olshausen, Bruno A., and David J. Field. "Emergence of simple-cell receptive field properties by learning a sparse code for natural images." Nature 381.6583 (1996): 607.
}
