% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fracExp}
\alias{fracExp}
\title{Calculate fractional exponent/power}
\usage{
fracExp(a, num, den)
}
\arguments{
\item{a}{numeric, base (could be nagetive).}

\item{num}{positive integer, numerator of the exponent.}

\item{den}{positive integer, denominator of the exponent.}
}
\value{
numeric, the evaluated a^(num/den)
}
\description{
Calculate fractional exponent/power, a^(num/den), where a could be negative.
}
