% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hard}
\alias{hard}
\title{Hard-thresholding}
\usage{
hard(x, t)
}
\arguments{
\item{x}{any numerical \code{matrix} or \code{vector}.}

\item{t}{\code{numeric}, the amount to hard-threshold, i.e., \eqn{sgn(x_{ij}) (|x_{ij}-t|)_+}.}
}
\description{
Perform hard-thresholding given the cut-off value.
}
