% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{inner}
\alias{inner}
\title{Custom inner produce of matrices}
\usage{
inner(X, Y, FUN = crossprod, ...)
}
\arguments{
\item{X, Y}{matrix or Matrix.}

\item{FUN}{function, which takes in two vectors and returns a numeric value.}

\item{...}{additional parameters for FUN.}
}
\value{
a matrix of <X, Y>
}
\description{
Calculate custom matrix inner product Z = <X, Y>, where Z[i,j] = FUN(X[,i], Y[,j]).
}
