% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{labelCluster}
\alias{labelCluster}
\title{Label cluster}
\usage{
labelCluster(x, ties.method = "random")
}
\arguments{
\item{x}{matrix, nNode x nBlock, non-negative, higher value means more likely}

\item{ties.method}{a character string specifying how ties are treated,
see \code{\link[base]{rank}}, only "random", "first", "last" are allowed.}
}
\value{
integer vector %in% 1:ncol(x)
}
\description{
Given the block membership matrix x, where x[i,j] is the likelihood of node i belongs to block j. Finds the cluster label.
}
