% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{matchCols}
\alias{matchCols}
\title{Match two matrices}
\usage{
matchCols(X.new, X, method = "maximum")
}
\arguments{
\item{X.new, X}{matrix or Matrix, with the same dimension.}

\item{method}{one of "maximum" (default), "cosine", or "euclidean".}
}
\value{
X.new after permuting the columns.
}
\description{
Match the columns of two matrices of the same size.
This is accomplished by solving an optimization over column permutation.
Given two matrices, X and X.new, find permutation p() that minimizes
     sum_i similarity(X.new[,p(i)], X[,i]),
where the similarity can be euclidean distance, cosine, maximum difference.
}
