% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{norm.Lp}
\alias{norm.Lp}
\title{Element-wise Matrix Norm}
\usage{
norm.Lp(mat, p = 1)
}
\arguments{
\item{mat}{a \code{matrix} or \code{Matrix}.}

\item{p}{\code{numeric(1)}, the p for defining the Lp norm.}
}
\value{
\code{numeric(1)}, the absolute sum of all elements.
}
\description{
Compute element-wise matrix Lp-norm.
This is a helper function to \code{\link[=shrinkage]{shrinkage()}}.
}
