% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{permColumn}
\alias{permColumn}
\title{Permute columns of a block membership matrix}
\usage{
permColumn(x, s = 2)
}
\arguments{
\item{x}{a non-negative matrix, nNode x nBlock,}

\item{s}{integer, order of non-linear}
}
\description{
Perform column permutation of block membership matrix for aesthetic visualization.
That is, the k-th column gives k-th cluster. This is done by ranking the column sums of squares (by default).
}
