% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epca.R
\docType{data}
\name{pitprops}
\alias{pitprops}
\title{Pitprops correlation data}
\description{
The \code{pitprops} data is a correlation matrix that was calculated from 180 observations.
There are 13 explanatory variables.
Jeffers (1967) tried to interpret the first six PCs.
This is a classical example showing the difficulty of interpreting principal components.
}
\examples{
\donttest{
## NOT TEST
data(pitprops)
ggcorrplot::ggcorrplot(pitprops)
}

}
\references{
Jeffers, J. (1967) "Two case studies in the application of principal component", \emph{Applied Statistics}, 16, 225-236.
}
\keyword{datasets}
