% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{polar}
\alias{polar}
\title{Polar decomposition}
\usage{
polar(X)
}
\arguments{
\item{X}{matrix or Matrix.}
}
\value{
a matrix of the U part
}
\description{
Takes a matrix X, of dimensions n by p, and calculate polar components,
two matrices P and Z, such that X = UP,
assuming X has full col rank
}
