% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prewhiten}
\alias{prewhiten}
\title{Pre-whiten Image}
\usage{
prewhiten(img)
}
\arguments{
\item{img}{a \code{matrix} of square image.}
}
\description{
Mimic/simulate the processing in retina, the lateral geniculate nucleus (LGN), and V1
(1) scale all pixels to constant variance (e.g., 1).
(2) combined low-pass/whitening filter
}
\references{
Olshausen, Bruno A., and David J. Field. "Emergence of simple-cell receptive field properties by learning a sparse code for natural images." Nature 381.6583 (1996): 607.
}
