% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.sca}
\alias{print.sca}
\title{Print SCA}
\usage{
\method{print}{sca}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an \code{sca} object.}

\item{verbose}{\code{logical(1)}, whether to print out loadings.}

\item{...}{additional input to generic \link{print}.}
}
\value{
Print an \code{sma} object interactively.
}
\description{
Print SCA
}
