% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.sma}
\alias{print.sma}
\title{Print SMA}
\usage{
\method{print}{sma}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an \code{sma} object.}

\item{verbose}{\code{logical(1)}, whether to print out loadings.}

\item{...}{additional input to generic \link{print}.}
}
\value{
Print an \code{sma} object interactively.
}
\description{
Print SMA
}
