% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{prs}
\alias{prs}
\title{Polar-Rotation-Shrinkage}
\usage{
prs(X, Z.hat, gamma, rotate, shrink, normalize, order, flip, epsilon = 1e-05)
}
\arguments{
\item{X, Z.hat}{the matrix product \code{crossprod(X, Z.hat)} is the input.}

\item{gamma}{\code{numeric}, the sparsity parameter.}

\item{rotate}{\code{character(1)}, rotation method, "varimax" (default) or "absmin".}

\item{shrink}{\code{character(1)}, shrinkage method, "soft" (default) or "hard".}

\item{normalize}{\code{logical}, whether to normalize rows before rotation, then scale back after.}

\item{order}{\code{logical}, whether to re-order decreasingly the columns of estimates by PVE.}

\item{flip}{\code{logical}, whether to flip the signs of the columns of estimates such that all columns are positive-skewed, i.e., the sums of cubic elements are positive.}

\item{epsilon}{\code{numeric}, tolerance of convergence precision (dafault to 0.00001).}
}
\value{
a \code{matrix} of the same dimension as \code{A}, sparse estimate.
}
\description{
This function is a helper function of \code{\link[=sma]{sma()}}.
It performs polar docomposition, orthgonal rotation, and soft-thresholding shrinkage in order.
The three steps together enable sparse estimates of the SMA and SCA.
}
