% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rootmatrix}
\alias{rootmatrix}
\title{Find root matrix}
\usage{
rootmatrix(x)
}
\arguments{
\item{x}{a symmetric \code{matrix} (will trigger error if not symmetric).}
}
\description{
Find the root matrix (\code{x}) from the Gram matrix (i.e., \code{crossprod(x)}).
This is also useful when the input is a covariance matrix, up to a scaling factor of n-1, where n is the sample size.
}
