% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rotation}
\alias{rotation}
\title{Varimax Rotation}
\usage{
rotation(x, rotate = "varimax", normalize = F, flip = T, eps = 1e-06)
}
\arguments{
\item{x}{a \code{matrix} or \code{Matrix}.}

\item{rotate}{\code{character(1)}, rotation method, "varimax" (default) or "absmin".}

\item{normalize}{\code{logical}, whether to (default to \code{FALSE}) first normalize rows before (and scale back afterward) the rotation.}

\item{eps}{\code{numeric} pricision tolerance.}
}
\value{
the rotated matrix of the same dimension as \code{x}.
}
\description{
Perform varimax rotation.
Flip the signs of columns so that the resulting matrix is positive-skewed.
}
