% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sca}
\alias{sca}
\title{Sparse component analysis}
\usage{
sca(
  A,
  k = min(5, dim(A)),
  gamma = NULL,
  is.cov = F,
  rotate = "varimax",
  shrink = "soft",
  center = T,
  scale = T,
  normalize = F,
  order = T,
  flip = T,
  max.iter = 1000,
  epsilon = 1e-05,
  quiet = T
)
}
\arguments{
\item{A}{\code{matrix} or \code{Matrix} to be analyzed.}

\item{k}{\code{integer}, rank of approximation.}

\item{gamma}{\code{numeric(1)}, sparsity parameter, default to √(pk), where n x p is the dimension of \code{A}.}

\item{is.cov}{\code{logical}, whether the \code{A} is a covariance matrix or Gram matrix (i.e., \code{crossprod(X)}). This function presumes that \code{A} is \emph{not} covariance matrix.}

\item{rotate}{\code{character(1)}, rotation method, "varimax" (default) or "absmin".}

\item{shrink}{\code{character(1)}, shrinkage method, "soft" (default) or "hard".}

\item{center}{\code{logical}, whether to center columns of \code{A} (default to \code{TRUE}).}

\item{scale}{\code{logical}, whether to scale columns of \code{A} (default to \code{TRUE}).}

\item{normalize}{\code{logical}, whether to normalize rows before rotation, then scale back after.}

\item{order}{\code{logical}, whether to re-order decreasingly the columns of estimates by PVE.}

\item{flip}{\code{logical}, whether to flip the signs of the columns of estimates such that all columns are positive-skewed, i.e., the sums of cubic elements are positive.}

\item{max.iter}{\code{integer}, maximum number of iteration (default to 1,000).}

\item{epsilon}{\code{numeric}, tolerance of convergence precision (dafault to 0.00001).}

\item{quiet}{\code{logical}, whether to mute the process report (default to \code{TRUE})}
}
\value{
a list of
\item{loadings}{\code{matrix}, sparse loadings of PCs.}
\item{scores}{an n x k \code{matrix}, the component scores.}
\item{sdev}{a \code{numeric} vector of length \code{k}, standard deviation of each columns of scores. These may not sum to exactly 1 because of a slight loss of orthogonality.}
\item{pve}{a \code{numeric} vector of length \code{k}, cumulative proportion of variance in \code{A} explained by the top PCs.}
\item{center}{\code{logical}, this records the \code{center} parameter.}
\item{scale}{\code{logical}, this records the \code{scale} parameter.}
\item{n.iter}{\code{integer}, number of iteration tabke.}
\item{n.obs}{\code{integer}, sample size, that is, \code{nrow(A)}.}
}
\description{
Perform sparse component analysis (SCA).
SCA is a method of sparse PCA where the loadings are column sparse.
}
