% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sma}
\alias{sma}
\title{Sparse matrix approximation}
\usage{
sma(
  A,
  k = min(5, dim(A)),
  gamma = NULL,
  rotate = "varimax",
  shrink = "soft",
  center = F,
  scale = F,
  normalize = F,
  order = F,
  flip = F,
  max.iter = 1000,
  epsilon = 1e-05,
  quiet = T
)
}
\arguments{
\item{A}{\code{matrix} or \code{Matrix} to be analyzed.}

\item{k}{\code{integer}, rank of approximation.}

\item{gamma}{\code{numeric(2)}, sparsity parameters. If \code{gamma} is \code{numeric(1)}, it is used for both left and right sparsity component (i.e, Z and Y). If absent, the two parameters are set as (default): √(nk) and √(pk) for Z and Y respoectively, where n x p is the dimension of \code{A}.}

\item{rotate}{\code{character(1)}, rotation method, "varimax" (default) or "absmin".}

\item{shrink}{\code{character(1)}, shrinkage method, "soft" (default) or "hard".}

\item{center}{\code{logical}, whether to center columns of \code{A} (default to \code{TRUE}).}

\item{scale}{\code{logical}, whether to scale columns of \code{A} (default to \code{TRUE}).}

\item{normalize}{\code{logical}, whether to normalize rows before rotation, then scale back after.}

\item{order}{\code{logical}, whether to re-order decreasingly the columns of estimates by PVE.}

\item{flip}{\code{logical}, whether to flip the signs of the columns of estimates such that all columns are positive-skewed, i.e., the sums of cubic elements are positive.}

\item{max.iter}{\code{integer}, maximum number of iteration (default to 1,000).}

\item{epsilon}{\code{numeric}, tolerance of convergence precision (dafault to 0.00001).}

\item{quiet}{\code{logical}, whether to mute the process report (default to \code{TRUE})}
}
\value{
a \code{list} of four elements:
\item{Z}{an n x k \code{matrix} , sparse component Z of \code{A}'srow space.}
\item{B}{a k x k \code{matrix}, the middle B matrix. This is like the scores of the SMA, because the objective of SMA is to maximize the Frobenius norm of B.}
\item{Y}{an n x k \code{matrix}, sparse component Y of \code{A}'s column space.}
\item{score}{optimal objective value.}
\item{n.iter}{an \code{integer}, the number of iteration taken.}
}
\description{
Perform the sparse matrix approximation (SMA) of a data matrix X as three components: Z B Y',
where Z and Y are sparse, and B is low-rank but not necessarily diagonal.
}
