% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\title{Sparse multivariate decomposition (SMD)}
\usage{
smd(A, k = 5, lambda = NULL, side = c("both", "left", "right"),
  centering = F, scaling = F, normalizing = F, max.iter = 1000,
  epsilon = 1e-05, quiet = T)
}
\arguments{
\item{A}{matrix or Matrix to be decomposed, of dimensions n x p}

\item{k}{integer, rank of approximation}

\item{lambda}{numeric, penalized parameter, default is √n√k (for Z) and/or √p√k (for Y)}

\item{centering}{logical, whether to (TRUE) center columns of A}

\item{scaling}{logical, whether to (TRUE) scale columns of A}

\item{normalizing}{logical, whether to (TRUE) normalize rows before rotation, then scale back after}

\item{epsilon}{numeric, tolerance of precision}
}
\value{
a list() containing
$Z - matrix n x k, sparse left singular vector (in column)
$B - matrix k x k, Z^T A Y
$Y - matrix n x k, sparse right singluar vector (in column)
$n.iter - integer, number of iteration taken
}
\description{
Perform sparse multivariate decomposition (SMD): A = Z B Y^T
}
