% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{sparse.pca}
\alias{sparse.pca}
\title{Sparse PCA via SMD}
\usage{
sparse.pca(A, k = 5, lambda = NULL, input.type = "predictor",
  centering = T, scaling = F, normalizing = F, ordering = T,
  max.iter = 1000, epsilon = 1e-05, quiet = T)
}
\arguments{
\item{A}{matrix (or Matrix), data matrix}

\item{k}{number of PCs to compute}

\item{lambda}{penalized parameter, default is sqrt(nrow(A)) and sqrt(ncol(A))}

\item{input.type}{'predictor' (original) or 'Gram' (covariance, when n >> p), the input matrix for SMD}

\item{centering}{logical, whether to (TRUE) center columns of A}

\item{scaling}{logical, whether to (TRUE) scale columns of A}

\item{normalizing}{logical, whether to (TRUE) normalize rows before rotation, then scale back after}

\item{ordering}{logical or "sdev" or "pve", whether to (TRUE) reorder columns of loadings by decreasing P.V.E.}

\item{epsilon}{tolerance}
}
\value{
a list of
   $sdev - numeric vector, standard deviation of PC,
   $loadings - (sparse) matrix, loadings of PC,
   $center - logical (=centering),
   $scale - logical (=scaling),
   $n.obs - integer, sample size, = nrow(A),
   $n.iter - integer, number of iterations taken,
   $scores - matrix of n*k, the scores of the supplied data on the principal components.
}
\description{
Perform sparse PCA via the sparse multivariate decomposition (SMD)
}
