% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{trim.fringe}
\alias{trim.fringe}
\title{Trim Image}
\usage{
trim.fringe(img, hem = 6)
}
\arguments{
\item{img}{a \code{matrix} of square image.}

\item{hem}{\code{integer}, the width of trimmed fringes.}
}
\description{
Trim the fringes of image
}
