% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{var.explain}
\alias{var.explain}
\title{Proportion of variance explained}
\usage{
var.explain(mat, V, type = "predictor")
}
\arguments{
\item{mat}{matrix or Matrix, the original data matrix X or cov(X) = crossprod(X) / (nrow(X) - 1)}

\item{V}{coefficients of linear transformation, e.g. a set of loadings in PCA.}

\item{type}{the input matrix for SMD (not used at the moment), either
'predictor' (original data matrix) or 'Gram' (useful when n >> p).}
}
\value{
a numeric value between [0,1], the proportion of variance in mat explained by Y.
}
\description{
Calculate the variance in a matrix explained by a set of linear transformation, (e.g. eigenvectors).
}
