% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{varimax.criteria}
\alias{varimax.criteria}
\title{The varimax criteria}
\usage{
varimax.criteria(mat)
}
\description{
Calculate the varimax criteria.
}
\references{
\href{https://en.wikipedia.org/wiki/Varimax_rotation}{Varimax rotation (Wikipedia)}
}
