% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{varimax}
\alias{varimax}
\title{Re-implementation of varimax
This is an extended version of \code{\link[stats]{varimax}}, allowing control of iteration.}
\usage{
varimax(x, normalize = F, eps = 1e-05, maxit = 1000L)
}
\arguments{
\item{x}{A loadings matrix, with \eqn{p} rows and \eqn{k < p} columns}

\item{normalize}{logical. Should Kaiser normalization be performed?
    If so the rows of \code{x} are re-scaled to unit length before
    rotation, and scaled back afterwards.}

\item{eps}{The tolerance for stopping: the relative change in the sum
    of singular values.}
}
\description{
Re-implementation of varimax
This is an extended version of \code{\link[stats]{varimax}}, allowing control of iteration.
}
