% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{absmin.criteria}
\alias{absmin.criteria}
\title{Absmin Criteria}
\usage{
absmin.criteria(L)
}
\arguments{
\item{L}{a \code{matrix} or \code{Matrix}.}
}
\description{
Calculate the absmin criteria.
This is a helper function for \link{absmin}.
}
