% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epca.R
\docType{package}
\name{epca-package}
\alias{epca-package}
\title{Exploratory Principal Component Analysis}
\description{
\code{epca} is for comprehending any data matrix that contains \emph{low-rank} and \emph{sparse} underlying signals of interest.
The package currently features two key tools: (1) \code{sca} for \strong{s}parse principal \strong{c}omponent \strong{a}nalysis and
(2) \code{sma} for \strong{s}parse \strong{m}atrix \strong{a}pproximation, a two-way data analysis for simultaneously row and column dimensionality reductions.
}
\references{
Chen, F. and Rohe K. (2020) "A New Basis for Sparse PCA".
}
\keyword{package}
