% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{exp.frac}
\alias{exp.frac}
\title{Calculate fractional exponent/power}
\usage{
\method{exp}{frac}(a, num, den)
}
\arguments{
\item{a}{\code{numeric(1)}, base (could be negative).}

\item{num}{a positive \code{integer}, numerator of the exponent.}

\item{den}{a positive \code{integer}, denominator of the exponent.}
}
\value{
\code{numeric}, the evaluated a^(num/den)
}
\description{
Calculate fractional exponent/power, \code{a^(num/den)}, where a could be negative.
}
