% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{inner}
\alias{inner}
\title{Matrix Inner Product}
\usage{
inner(X, Y, FUN = "crossprod", ...)
}
\arguments{
\item{X, Y}{\code{matrix} or \code{Matrix}.}

\item{FUN}{\code{function} or a \code{character(1)} name of base function. The function should take in two vectors as input and ouput a \code{numeric(1)} result.}

\item{...}{additional parameters for \code{FUN}.}
}
\value{
\code{matrix}, inner product of \code{X} and \code{Y}.
}
\description{
Calculate the custom matrix inner product, Z = <X, Y>, where Z\link{i,j} = FUN(X\link{,i}, Y\link{,j}).
}
