% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{labelCluster}
\alias{labelCluster}
\title{Label Cluster}
\usage{
labelCluster(x, ties.method = "random")
}
\arguments{
\item{x}{\code{matrix} with non-negative entries, where \code{x[i,j]} is the estimated likelihood (or any equivalent measure) of node i belongs to block j. The higher the more likely.}

\item{ties.method}{\code{character}, how should ties be handled, "random", "first", "last" are allowed. See \code{\link[base:rank]{base::rank()}} for details.}
}
\value{
\code{integer} vector of the same length as \code{x}. Each entry is one of 1, 2, ..., \code{ncol(x)}.
}
\description{
Assign cluster labels to each row from the membership matrix.
}
