% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rootmatrix}
\alias{rootmatrix}
\title{Find root matrix}
\usage{
rootmatrix(x)
}
\arguments{
\item{x}{a symmetric \code{matrix}.}
}
\description{
Find \code{X} from the Gram matrix (i.e., \code{crossprod(X)}).
}
