% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rotation}
\alias{rotation}
\title{Orthogonal rotation for the SMD}
\usage{
rotation(X, normalize = F, eps = 1e-06)
}
\arguments{
\item{X}{matrix or Matrix.}

\item{normalize}{logical, whether to (default to FALSE) first normalize rows before (and scale back afterward) the rotation.}

\item{eps}{pricision tolerance.}
}
\value{
the rotated matrix
}
\description{
Perform varimax rotation on X from right.
Flip the signs of columns so that the resulting matrix is positive skewed.
}
