% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\title{Sparse multivariate decomposition (SMD)}
\usage{
smd(A, k = 5, lambda = NULL, side = c("both", "left", "right"),
  centering = F, scaling = F, normalizing = F, max.iter = 1000,
  epsilon = 1e-05, quiet = T)
}
\arguments{
\item{A}{matrix or Matrix to be decomposed, of dimensions n x p}

\item{k}{integer, rank of approximation}

\item{lambda}{numeric, penalized parameter, default is √(nk) (for Z) and/or √(pk) (for Y)}

\item{centering}{logical, whether to (default to `TRUE`) center columns of A}

\item{scaling}{logical, whether to (default to `TRUE`) scale columns of A}

\item{normalizing}{logical, whether to (default to `TRUE`) normalize rows before rotation, then scale back after}

\item{epsilon}{numeric, tolerance of precision}
}
\value{
a list
\item{Z}{an n x k matrix , sparse left singular vector (in column).}
\item{B}{a k x k matrix, calculated by Z' A Y.}
\item{Y}{an n x k matrix, sparse right singluar vector (in column).}
\item{n.iter}{an integer, number of iteration taken.}
}
\description{
Perform sparse multivariate decomposition (SMD): A = Z B Y', where Z and Y are sparse, and B is low-rank.
}
\examples{

}
