% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{soft}
\alias{soft}
\title{Soft-thresholding}
\usage{
soft(x, t)
}
\arguments{
\item{x}{any numerical \code{matrix} or \code{vector}.}

\item{t}{\code{numeric}, the amount to soft-threshold, i.e., \eqn{sgn(x_{ij}) (|x_{ij}-t|)_+}.}
}
\description{
Perform soft-thresholding given the cut-off value.
}
