% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{vgQ.absmin}
\alias{vgQ.absmin}
\title{Gradient of Absmin Criterion}
\usage{
vgQ.absmin(L)
}
\arguments{
\item{L}{a \code{matrix} or \code{Matrix}.}
}
\value{
a list required by \code{GPArotation::GPForth} for the absmin rotation.
}
\description{
This is a helper function for \link{absmin} and is not to be used directly by users.
}
\examples{
\dontrun{
## NOT RUN
## NOT for users to call.
}
}
